refactor: Extract element builder for parameterized tests
The builder is useful for parameterized tests, for example, to run a test case for different BPMN elements.
